+++ gradio/blocks.py
@@ -24,7 +24,7 @@ from gradio.context import Context
 from gradio.deprecation import check_deprecated_parameters
 from gradio.documentation import document, set_documentation_group
 from gradio.exceptions import DuplicateBlockError, InvalidApiName
-from gradio.helpers import create_tracker, skip, special_args
+from gradio.helpers import EventData, create_tracker, skip, special_args
 from gradio.themes import Default as DefaultTheme
 from gradio.themes import ThemeClass as Theme
 from gradio.tunneling import CURRENT_TUNNELS
@@ -140,6 +140,7 @@ class Block:
         max_batch_size: int = 4,
         cancels: List[int] | None = None,
         every: float | None = None,
+        collects_event_data: bool | None = None,
         trigger_after: int | None = None,
         trigger_only_on_success: bool = False,
     ) -> Tuple[Dict[str, Any], int]:
@@ -161,6 +162,7 @@ class Block:
             max_batch_size: the maximum batch size to send to the function
             cancels: a list of other events to cancel when this event is triggered. For example, setting cancels=[click_event] will cancel the click_event, where click_event is the return value of another components .click method.
             every: Run this event 'every' number of seconds while the client connection is open. Interpreted in seconds. Queue must be enabled.
+            collects_event_data: whether to collect event data for this event
             trigger_after: if set, this event will be triggered after 'trigger_after' function index
             trigger_only_on_success: if True, this event will only be triggered if the previous event was successful (only applies if `trigger_after` is set)
         Returns: dependency information, dependency index
@@ -204,8 +206,19 @@ class Block:
         elif every:
             raise ValueError("Cannot set a value for `every` without a `fn`.")
 
+        _, progress_index, event_data_index = (
+            special_args(fn) if fn else (None, None, None)
+        )
         Context.root_block.fns.append(
-            BlockFunction(fn, inputs, outputs, preprocess, postprocess, inputs_as_dict)
+            BlockFunction(
+                fn,
+                inputs,
+                outputs,
+                preprocess,
+                postprocess,
+                inputs_as_dict,
+                progress_index is not None,
+            )
         )
         if api_name is not None:
             api_name_ = utils.append_unique_suffix(
@@ -216,6 +229,9 @@ class Block:
                     "api_name {} already exists, using {}".format(api_name, api_name_)
                 )
                 api_name = api_name_
+
+        if collects_event_data is None:
+            collects_event_data = event_data_index is not None
 
         dependency = {
             "targets": [self._id] if not no_target else [],
@@ -236,6 +252,7 @@ class Block:
                 "continuous": bool(every),
                 "generator": inspect.isgeneratorfunction(fn) or bool(every),
             },
+            "collects_event_data": collects_event_data,
             "trigger_after": trigger_after,
             "trigger_only_on_success": trigger_only_on_success,
         }
@@ -276,7 +293,7 @@ class BlockContext(Block):
             render: If False, this will not be included in the Blocks config file at all.
         """
         self.children: List[Block] = []
-        super().__init__(visible=visible, render=render, **kwargs)
+        Block.__init__(self, visible=visible, render=render, **kwargs)
 
     def __enter__(self):
         self.parent = Context.block
@@ -330,12 +347,14 @@ class BlockFunction:
         preprocess: bool,
         postprocess: bool,
         inputs_as_dict: bool,
+        tracks_progress: bool = False,
     ):
         self.fn = fn
         self.inputs = inputs
         self.outputs = outputs
         self.preprocess = preprocess
         self.postprocess = postprocess
+        self.tracks_progress = tracks_progress
         self.total_runtime = 0
         self.total_runs = 0
         self.inputs_as_dict = inputs_as_dict
@@ -803,6 +822,7 @@ class Blocks(BlockContext):
         iterator: Iterator[Any] | None = None,
         requests: routes.Request | List[routes.Request] | None = None,
         event_id: str | None = None,
+        event_data: EventData | None = None,
     ):
         """
         Calls function with given index and preprocessed input, and measures process time.
@@ -812,6 +832,7 @@ class Blocks(BlockContext):
             iterator: iterator to use if function is a generator
             requests: requests to pass to function
             event_id: id of event in queue
+            event_data: data associated with event trigger
         """
         block_fn = self.fns[fn_index]
         assert block_fn.fn, f"function with index {fn_index} not defined."
@@ -829,10 +850,8 @@ class Blocks(BlockContext):
             request = requests[0]
         else:
             request = requests
-        processed_input, progress_index = special_args(
-            block_fn.fn,
-            processed_input,
-            request,
+        processed_input, progress_index, _ = special_args(
+            block_fn.fn, processed_input, request, event_data
         )
         progress_tracker = (
             processed_input[progress_index] if progress_index is not None else None
@@ -993,6 +1012,7 @@ class Blocks(BlockContext):
         request: routes.Request | List[routes.Request] | None = None,
         iterators: Dict[int, Any] | None = None,
         event_id: str | None = None,
+        event_data: EventData | None = None,
     ) -> Dict[str, Any]:
         """
         Processes API calls from the frontend. First preprocesses the data,
@@ -1003,6 +1023,8 @@ class Blocks(BlockContext):
             username: name of user if authentication is set up (not used)
             state: data stored from stateful components for session (key is input block id)
             iterators: the in-progress iterators for each generator function (key is function index)
+            event_id: id of event that triggered this API call
+            event_data: data associated with the event trigger itself
         Returns: None
         """
         block_fn = self.fns[fn_index]
@@ -1029,7 +1051,7 @@ class Blocks(BlockContext):
                 self.preprocess_data(fn_index, list(i), state) for i in zip(*inputs)
             ]
             result = await self.call_function(
-                fn_index, list(zip(*inputs)), None, request
+                fn_index, list(zip(*inputs)), None, request, event_id, event_data
             )
             preds = result["prediction"]
             data = [
@@ -1041,7 +1063,7 @@ class Blocks(BlockContext):
             inputs = self.preprocess_data(fn_index, inputs, state)
             iterator = iterators.get(fn_index, None) if iterators else None
             result = await self.call_function(
-                fn_index, inputs, iterator, request, event_id
+                fn_index, inputs, iterator, request, event_id, event_data
             )
             data = self.postprocess_data(fn_index, result["prediction"], state)
             is_generating, iterator = result["is_generating"], result["iterator"]
@@ -1124,10 +1146,7 @@ class Blocks(BlockContext):
             self.parent.children.extend(self.children)
         self.config = self.get_config_file()
         self.app = routes.App.create_app(self)
-        self.progress_tracking = any(
-            block_fn.fn is not None and special_args(block_fn.fn)[1] is not None
-            for block_fn in self.fns
-        )
+        self.progress_tracking = any(block_fn.tracks_progress for block_fn in self.fns)
 
     @class_or_instancemethod
     def load(