+++ test/test_events.py
@@ -1,4 +1,5 @@
 import pytest
+from fastapi.testclient import TestClient
 
 import gradio as gr
 
@@ -16,6 +17,26 @@ class TestEvent:
             img.clear(fn_img_cleared, [], [])
 
         assert demo.config["dependencies"][0]["trigger"] == "clear"
+
+    def test_event_data(self):
+        with gr.Blocks() as demo:
+            text = gr.Textbox()
+            gallery = gr.Gallery()
+
+            def fn_img_index(evt: gr.SelectData):
+                return evt.index
+
+            gallery.select(fn_img_index, None, text)
+
+        app, _, _ = demo.launch(prevent_thread_lock=True)
+        client = TestClient(app)
+
+        resp = client.post(
+            f"{demo.local_url}run/predict",
+            json={"fn_index": 0, "data": [], "event_data": {"index": 1, "value": None}},
+        )
+        assert resp.status_code == 200
+        assert resp.json()["data"][0] == "1"
 
     def test_consecutive_events(self):
         def double(x):