+++ test/test_routes.py
@@ -439,6 +439,27 @@ class TestPassingRequest:
         output = dict(response.json())
         assert output["data"] == ["test"]
 
+    def test_request_get_headers(self):
+        def identity(name, request: gr.Request):
+            assert isinstance(request.headers["user-agent"], str)
+            assert isinstance(request.headers.items(), list)
+            assert isinstance(request.headers.keys(), list)
+            assert isinstance(request.headers.values(), list)
+            assert isinstance(dict(request.headers), dict)
+            user_agent = request.headers["user-agent"]
+            assert "testclient" in user_agent
+            return name
+
+        app, _, _ = gr.Interface(identity, "textbox", "textbox").launch(
+            prevent_thread_lock=True,
+        )
+        client = TestClient(app)
+
+        response = client.post("/api/predict/", json={"data": ["test"]})
+        assert response.status_code == 200
+        output = dict(response.json())
+        assert output["data"] == ["test"]
+
     def test_request_includes_username_as_none_if_no_auth(self):
         def identity(name, request: gr.Request):
             assert request.username is None